THE MOST GODLY LINE EVAR +more
so i added "process.on('unhandledRejection', console.error);" and it
catches every error ever which is amazing.
lessened the amount of times maddie listens to stuff
makes sure maddie cant reply to herself
added blanket, triforce, and coffee, fixed northkinds
updated package.json for acurate author and start

+++ index.js
@@ -6,10 +6,12 @@ client.on('ready' ,() => {
   console.log('Online. The default global prefix is: ' + prefix) + '.';
   client.user.setGame('with Landen\'s coffee!');
 });
-
+process.on('unhandledRejection', console.error);
 var prefix = '!';
 
 client.on('message', message => {
+  if (!message.content.startsWith(prefix)) return;
+  if (message.author.bot) return;
   if (message.content === prefix + 'blanket') {
     message.delete()
     message.channel.send('http://prntscr.com/f7kpom');
